+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to MiniJinja are documented here.
 
 ## 0.31.0
 
-- Added the Jinja2 tests `==`, `!=`, `<`, `<=`, `>`, and `>=` for the
+- Added the Jinja2 tests `==`, `!=`, `<`, `<=`, `>`, `>=` and `in` for the
   use with `select` and `reject`.
 
 ## 0.30.2